Fix a bug where ZNC would fail to send out its own ip for DCC bouncing
CUtils::GetLongIP() had a bug where it would overflow an integer for IPs
which started which a number higher than 127.
This is now fixed by using unsigned integers and doing that shift on a long.


git-svn-id: https://znc.svn.sourceforge.net/svnroot/znc/trunk@1340 726aef4b-f618-498e-8847-2d620e286838

+++ Utils.cpp
@@ -149,17 +149,18 @@ CString CUtils::GetIP(unsigned long addr) {
 unsigned long CUtils::GetLongIP(const CString& sIP) {
 	unsigned long ret;
 	char ip[4][4];
-	int i;
+	unsigned int i;
 
 	i = sscanf(sIP.c_str(), "%3[0-9].%3[0-9].%3[0-9].%3[0-9]",
 			ip[0], ip[1], ip[2], ip[3]);
 	if (i != 4)
 		return 0;
 
-	ret  = atoi(ip[0]) << 24;
-	ret += atoi(ip[1]) << 16;
-	ret += atoi(ip[2]) << 8;
-	ret += atoi(ip[3]) << 0;
+	// Beware that atoi("200") << 24 would overflow and turn negative!
+	ret  = atol(ip[0]) << 24;
+	ret += atol(ip[1]) << 16;
+	ret += atol(ip[2]) << 8;
+	ret += atol(ip[3]) << 0;
 
 	return ret;
 }
